Update UI 	- Redid the UI
