+++ docker-compose.yml
@@ -46,6 +46,8 @@ services:
       dockerfile: ./src/app/Dockerfile
     ports:
       - "3000:3000"
+    environment:
+      - NEXT_PUBLIC_API_BASE=${NEXT_PUBLIC_API_BASE}
     volumes:
       - ./src/app:/recipe/app
       - /app/node_modules
