+++ src/app/components/RecipeCard.tsx
@@ -1,3 +1,5 @@
+import { getTagClassName } from "./tagColor";
+
 interface RecipeCardProps {
   recipe: {
     id: number;
@@ -6,40 +8,17 @@ interface RecipeCardProps {
     tags?: string[];
     image?: string;
   };
+  isFavorited: boolean;
+  onToggleFavorite: (recipeId: number) => void;
   onClick: () => void;
 }
 
-const tagStyles: Record<string, string> = {
-  appetizers: "bg-violet-100 text-violet-700",
-  dessert: "bg-orange-100 text-orange-700",
-  dinner: "bg-amber-100 text-amber-800",
-  lunch: "bg-[#417df6]/14 text-[#417df6]",
-  seafood: "bg-cyan-100 text-cyan-700",
-  vegan: "bg-emerald-100 text-emerald-700",
-  vegetarian: "bg-lime-100 text-lime-700",
-};
-
-const fallbackTagStyles = [
-  "bg-indigo-100 text-indigo-700",
-  "bg-rose-100 text-rose-700",
-  "bg-amber-100 text-amber-700",
-  "bg-teal-100 text-teal-700",
-];
-
-const getTagClassName = (tag: string) => {
-  const normalizedTag = tag.trim().toLowerCase();
-  if (tagStyles[normalizedTag]) {
-    return tagStyles[normalizedTag];
-  }
-
-  const charTotal = normalizedTag
-    .split("")
-    .reduce((total, char) => total + char.charCodeAt(0), 0);
-
-  return fallbackTagStyles[charTotal % fallbackTagStyles.length];
-};
-
-export default function RecipeCard({ recipe, onClick }: RecipeCardProps) {
+export default function RecipeCard({
+  recipe,
+  isFavorited,
+  onToggleFavorite,
+  onClick,
+}: RecipeCardProps) {
   return (
     <button
       onClick={onClick}
@@ -57,8 +36,30 @@ export default function RecipeCard({ recipe, onClick }: RecipeCardProps) {
             No Image
           </div>
         )}
-        <span className="absolute right-3 top-3 inline-flex h-8 w-8 items-center justify-center rounded-full bg-white/95 text-slate-500 shadow-[0_10px_24px_rgba(15,23,42,0.08)]">
-          <svg xmlns="http://www.w3.org/2000/svg" fill="none" viewBox="0 0 24 24" strokeWidth={1.8} stroke="currentColor" className="h-3.5 w-3.5">
+        <span
+          role="button"
+          tabIndex={0}
+          aria-label={isFavorited ? `Remove ${recipe.name} from favorites` : `Add ${recipe.name} to favorites`}
+          onClick={(event) => {
+            event.preventDefault();
+            event.stopPropagation();
+            onToggleFavorite(recipe.id);
+          }}
+          onKeyDown={(event) => {
+            if (event.key !== "Enter" && event.key !== " ") {
+              return;
+            }
+            event.preventDefault();
+            event.stopPropagation();
+            onToggleFavorite(recipe.id);
+          }}
+          className={`absolute right-3 top-3 inline-flex h-8 w-8 items-center justify-center rounded-full shadow-[0_10px_24px_rgba(15,23,42,0.08)] transition ${
+            isFavorited
+              ? "bg-rose-100 text-rose-500"
+              : "bg-white/95 text-slate-500 hover:bg-rose-50 hover:text-rose-500"
+          }`}
+        >
+          <svg xmlns="http://www.w3.org/2000/svg" fill={isFavorited ? "currentColor" : "none"} viewBox="0 0 24 24" strokeWidth={1.8} stroke="currentColor" className="h-3.5 w-3.5">
             <path strokeLinecap="round" strokeLinejoin="round" d="m21 8.25c0-2.485-2.015-4.5-4.5-4.5-1.74 0-3.25.988-4 2.433-.75-1.445-2.26-2.433-4-2.433-2.485 0-4.5 2.015-4.5 4.5 0 6.375 8.5 11.25 8.5 11.25S21 14.625 21 8.25Z" />
           </svg>
         </span>
@@ -79,7 +80,7 @@ export default function RecipeCard({ recipe, onClick }: RecipeCardProps) {
         <div className="mt-auto flex items-end gap-2 pt-4">
           {recipe.tags && recipe.tags.length > 0 && (
             <div className="flex flex-wrap items-center gap-1.5">
-              {recipe.tags.slice(0, 2).map((tag) => (
+              {recipe.tags.map((tag) => (
                 <span
                   key={tag}
                   className={`rounded-full px-2.5 py-0.5 text-[10px] font-medium ${getTagClassName(tag)}`}
@@ -87,11 +88,6 @@ export default function RecipeCard({ recipe, onClick }: RecipeCardProps) {
                   {tag}
                 </span>
               ))}
-              {recipe.tags.length > 2 && (
-                <span className="rounded-full bg-slate-100 px-2 py-0.5 text-[10px] font-medium text-slate-500">
-                  +{recipe.tags.length - 2}
-                </span>
-              )}
             </div>
           )}
         </div>
